Pass end position of span through inline ASM cookie

+++ compiler/rustc_codegen_llvm/src/llvm/ffi.rs
@@ -2293,7 +2293,7 @@ extern "C" {
 
     pub fn LLVMRustGetSMDiagnostic<'a>(
         DI: &'a DiagnosticInfo,
-        cookie_out: &mut c_uint,
+        cookie_out: &mut u64,
     ) -> &'a SMDiagnostic;
 
     #[allow(improper_ctypes)]
